+++ sdmx-dl-api/src/main/java/sdmxdl/HasDescription.java
@@ -1,0 +1,13 @@
+package sdmxdl;
+
+import lombok.NonNull;
+
+public interface HasDescription {
+
+    /**
+     * Gets a human-readable (best-language-match) description.
+     *
+     * @return a non-null description
+     */
+    @NonNull String getDescription();
+}