+++ sdmx-dl-api/src/main/java/sdmxdl/HasName.java
@@ -1,0 +1,13 @@
+package sdmxdl;
+
+import lombok.NonNull;
+
+public interface HasName {
+
+    /**
+     * Gets a human-readable (best-language-match) name.
+     *
+     * @return a non-null name
+     */
+    @NonNull String getName();
+}