+++ sdmx-dl-api/src/test/java/sdmxdl/DataRepositoryTest.java
@@ -71,7 +71,7 @@ public class DataRepositoryTest {
     private final DataflowRef goodFlowRef = DataflowRef.of("NBB", "XYZ", "v2.0");
     private final DataflowRef badFlowRef = DataflowRef.parse("other");
     private final Dataflow flow = Dataflow.builder().ref(goodFlowRef).structureRef(goodStructRef).name("flow1").build();
-    private final DataStructure struct = DataStructure.builder().ref(goodStructRef).primaryMeasureId("").label("struct1").build();
+    private final DataStructure struct = DataStructure.builder().ref(goodStructRef).primaryMeasureId("").name("struct1").build();
     private final Obs obs1 = Obs.builder().period(LocalDateTime.now()).value(Math.PI).build();
     private final Series series = Series.builder().key(Key.of("BE")).obs(obs1).meta("hello", "world").build();
     private final DataSet dataSet = DataSet.builder().ref(goodFlowRef).series(series).build();