+++ sdmx-dl-api/src/test/java/tests/sdmxdl/api/RepoSamples.java
@@ -49,12 +49,12 @@ public class RepoSamples {
     public static final Codelist CL3 = Codelist.builder().ref(CL_REF3).code("INDUSTRY", "Industry").code("XXX", "Other").build();
     public static final Codelist CL4 = Codelist.builder().ref(CL_REF4).code("A", "Normal value").build();
 
-    public static final Dimension DIM1 = Dimension.builder().id("FREQ").codelist(CL1).label("Frequency").position(1).build();
-    public static final Dimension DIM2 = Dimension.builder().id("REGION").codelist(CL2).label("Region").position(3).build();
-    public static final Dimension DIM3 = Dimension.builder().id("SECTOR").codelist(CL3).label("Sector").position(4).build();
+    public static final Dimension DIM1 = Dimension.builder().id("FREQ").codelist(CL1).name("Frequency").position(1).build();
+    public static final Dimension DIM2 = Dimension.builder().id("REGION").codelist(CL2).name("Region").position(3).build();
+    public static final Dimension DIM3 = Dimension.builder().id("SECTOR").codelist(CL3).name("Sector").position(4).build();
 
-    public static final Attribute NOT_CODED_ATTRIBUTE = Attribute.builder().id("TITLE").label("Title").relationship(AttributeRelationship.SERIES).build();
-    public static final Attribute CODED_ATTRIBUTE = Attribute.builder().id("OBS_STATUS").codelist(CL4).label("Observation status").relationship(AttributeRelationship.OBSERVATION).build();
+    public static final Attribute NOT_CODED_ATTRIBUTE = Attribute.builder().id("TITLE").name("Title").relationship(AttributeRelationship.SERIES).build();
+    public static final Attribute CODED_ATTRIBUTE = Attribute.builder().id("OBS_STATUS").codelist(CL4).name("Observation status").relationship(AttributeRelationship.OBSERVATION).build();
 
     public static final DataStructure STRUCT = DataStructure
             .builder()
@@ -66,7 +66,7 @@ public class RepoSamples {
             .attribute(CODED_ATTRIBUTE)
             .timeDimensionId("TIME")
             .primaryMeasureId("OBS_VALUE")
-            .label("structName")
+            .name("structName")
             .build();
 
     public static final Obs OBS1 = Obs.builder().period(dateTimeOf(2010, 1)).value(Math.PI).build();