+++ sdmx-dl-provider-connectors/src/test/java/_test/sdmxdl/connectors/samples/ParsersTest.java
@@ -75,7 +75,7 @@ public class ParsersTest {
                                 .satisfies(dimension -> {
                                     assertThat(dimension.getId())
                                             .isEqualTo("FREQ");
-                                    assertThat(dimension.getLabel())
+                                    assertThat(dimension.getName())
                                             .isEqualTo("Frequency");
                                     assertThat(dimension.getPosition())
                                             .isEqualTo(1);
@@ -93,7 +93,7 @@ public class ParsersTest {
                                 .satisfies(attribute -> {
                                     assertThat(attribute.getId())
                                             .isEqualTo("OBS_STATUS");
-                                    assertThat(attribute.getLabel())
+                                    assertThat(attribute.getName())
                                             .isEqualTo("Observation status");
                                     assertThat(attribute.getCodelist()).satisfies(codelist -> {
                                         assertThat(codelist.getRef())
@@ -107,7 +107,7 @@ public class ParsersTest {
                                 .isEqualTo("TIME_PERIOD");
                         assertThat(dsd.getPrimaryMeasureId())
                                 .isEqualTo("OBS_VALUE");
-                        assertThat(dsd.getLabel())
+                        assertThat(dsd.getName())
                                 .isEqualTo("AMECO");
                     });
 
@@ -139,7 +139,7 @@ public class ParsersTest {
                                 .satisfies(dimension -> {
                                     assertThat(dimension.getId())
                                             .isEqualTo("SUBJECT");
-                                    assertThat(dimension.getLabel())
+                                    assertThat(dimension.getName())
                                             .isEqualTo("Sujet");
                                     assertThat(dimension.getPosition())
                                             .isEqualTo(1);
@@ -157,7 +157,7 @@ public class ParsersTest {
                                 .satisfies(attribute -> {
                                     assertThat(attribute.getId())
                                             .isEqualTo("OBS_STATUS");
-                                    assertThat(attribute.getLabel())
+                                    assertThat(attribute.getName())
                                             .isEqualTo("Observation Status");
                                     assertThat(attribute.getCodelist()).satisfies(codelist -> {
                                         assertThat(codelist.getRef())
@@ -171,7 +171,7 @@ public class ParsersTest {
                                 .isEqualTo("TIME");
                         assertThat(dsd.getPrimaryMeasureId())
                                 .isEqualTo("OBS_VALUE");
-                        assertThat(dsd.getLabel())
+                        assertThat(dsd.getName())
                                 .isEqualTo("Mon premier dataset");
                     });